requirements.txt
Django>=4.2,<5.0
celery>=5.3
redis>=5.0
selenium>=4.15
webdriver-manager>=4.0
gunicorn>=21.2
pyyaml>=6.0
